refactor: Move project structure and code structure analyzers
